// src/types.d.ts
export interface BoxInterface {
  height: number;
  width: number;
  left: number;
  top: number;
}

export interface Position {
  x: number;
  y: number;
}

export interface AreaInterface {
  top: number;
  right: number;
  bottom: number;
  left: number;
}

export interface BoxInfoInterface {
  width: number;
  height: number;
}